+++ src/object/transaction.cpp
@@ -33,7 +33,7 @@ transaction::transaction(const transaction &x)
   : transaction_data_(x.transaction_data_)
 {}
 
-unsigned long transaction::id() const
+unsigned long long transaction::id() const
 {
   return transaction_data_->id_;
 }
